requirements.txt
fastapi
uvicorn
flask
sqlalchemy
psycopg2-binary
openai
python-dotenv
requests
pytest
httpx
